Create empty pages - rename old pages to old.tsx
