+++ apps/web/app/(settings)/settings/security/sso/old.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+// This file has been sourced from: /Users/sean/Programming/cal.com/apps/web/pages/settings/security/sso.tsx
+import UserSSOView from "@calcom/features/ee/sso/page/user-sso-view";
+
+import type { CalPageWrapper } from "@components/PageWrapper";
+import PageWrapper from "@components/PageWrapper";
+
+const Page = UserSSOView as CalPageWrapper;
+Page.PageWrapper = PageWrapper;
+export default Page;
